+++ src/device-api/v1.ts
@@ -6,14 +6,19 @@ import { AuthorizedRequest } from './types';
 import * as eventTracker from '../event-tracker';
 import * as apiBinder from '../api-binder';
 import * as config from '../config';
+import * as hostConfig from '../host-config';
 import * as TargetState from '../device-state/target-state';
 import { getApp } from '../device-state/db-format';
+import * as deviceState from '../device-state';
 import * as applicationManager from '../compose/application-manager';
 import { generateStep } from '../compose/composition-steps';
 import * as commitStore from '../compose/commit';
 import { checkInt, checkTruthy } from '../lib/validation';
 import * as constants from '../lib/constants';
 import log from '../lib/supervisor-console';
+import { UpdatesLockedError } from '../lib/errors';
+
+const disallowedHostConfigPatchFields = ['local_ip', 'local_port'];
 
 export const router = express.Router();
 
@@ -118,6 +123,30 @@ const createStopOrStartHandler = (action: 'start' | 'stop') =>
 router.post('/v1/apps/:appId/stop', createStopOrStartHandler('stop'));
 router.post('/v1/apps/:appId/start', createStopOrStartHandler('start'));
 
+const rebootOrShutdown = async (
+	req: express.Request,
+	res: express.Response,
+	action: deviceState.DeviceStateStepTarget,
+) => {
+	const override = await config.get('lockOverride');
+	const force = checkTruthy(req.body.force) || override;
+	try {
+		const response = await deviceState.executeStepAction({ action }, { force });
+		res.status(202).json(response);
+	} catch (e) {
+		const status = e instanceof UpdatesLockedError ? 423 : 500;
+		res.status(status).json({
+			Data: '',
+			Error:
+				(e != null ? (e as Error).message : undefined) || e || 'Unknown error',
+		});
+	}
+};
+router.post('/v1/reboot', (req, res) => rebootOrShutdown(req, res, 'reboot'));
+router.post('/v1/shutdown', (req, res) =>
+	rebootOrShutdown(req, res, 'shutdown'),
+);
+
 router.get('/v1/apps/:appId', async (req: AuthorizedRequest, res, next) => {
 	const appId = checkInt(req.params.appId);
 	eventTracker.track('GET app (v1)', { appId });
@@ -211,3 +240,117 @@ router.post('/v1/update', (req: AuthorizedRequest, res, next) => {
 		res.sendStatus(202);
 	}
 });
+
+router.get('/v1/device/host-config', (_req, res) =>
+	hostConfig
+		.get()
+		.then((conf) => res.json(conf))
+		.catch((err) =>
+			res.status(503).send(err?.message ?? err ?? 'Unknown error'),
+		),
+);
+
+router.patch('/v1/device/host-config', async (req, res) => {
+	// Because v1 endpoints are legacy, and this endpoint might already be used
+	// by multiple users, adding too many throws might have unintended side effects.
+	// Thus we're simply logging invalid fields and allowing the request to continue.
+
+	try {
+		if (!req.body.network) {
+			log.warn("Key 'network' must exist in PATCH body");
+			// If network does not exist, skip all field validation checks below
+			throw new Error();
+		}
+
+		const { proxy } = req.body.network;
+
+		// Validate proxy fields, if they exist
+		if (proxy && Object.keys(proxy).length) {
+			const blacklistedFields = Object.keys(proxy).filter((key) =>
+				disallowedHostConfigPatchFields.includes(key),
+			);
+
+			if (blacklistedFields.length > 0) {
+				log.warn(`Invalid proxy field(s): ${blacklistedFields.join(', ')}`);
+			}
+
+			if (
+				proxy.type &&
+				!constants.validRedsocksProxyTypes.includes(proxy.type)
+			) {
+				log.warn(
+					`Invalid redsocks proxy type, must be one of ${constants.validRedsocksProxyTypes.join(
+						', ',
+					)}`,
+				);
+			}
+
+			if (proxy.noProxy && !Array.isArray(proxy.noProxy)) {
+				log.warn('noProxy field must be an array of addresses');
+			}
+		}
+	} catch (e) {
+		/* noop */
+	}
+
+	try {
+		// If hostname is an empty string, return first 7 digits of device uuid
+		if (req.body.network?.hostname === '') {
+			const uuid = await config.get('uuid');
+			req.body.network.hostname = uuid?.slice(0, 7);
+		}
+		const lockOverride = await config.get('lockOverride');
+		await hostConfig.patch(
+			req.body,
+			checkTruthy(req.body.force) || lockOverride,
+		);
+		res.status(200).send('OK');
+	} catch (err) {
+		// TODO: We should be able to throw err if it's UpdatesLockedError
+		// and the error middleware will handle it, but this doesn't work in
+		// the test environment. Fix this when fixing API tests.
+		if (err instanceof UpdatesLockedError) {
+			return res.status(423).send(err?.message ?? err);
+		}
+		res.status(503).send((err as Error)?.message ?? err ?? 'Unknown error');
+	}
+});
+
+router.get('/v1/device', async (_req, res) => {
+	try {
+		const state = await deviceState.getLegacyState();
+		const stateToSend = _.pick(state.local, [
+			'api_port',
+			'ip_address',
+			'os_version',
+			'mac_address',
+			'supervisor_version',
+			'update_pending',
+			'update_failed',
+			'update_downloaded',
+		]) as Dictionary<unknown>;
+		if (state.local?.is_on__commit != null) {
+			stateToSend.commit = state.local.is_on__commit;
+		}
+		const service = _.toPairs(
+			_.toPairs(state.local?.apps)[0]?.[1]?.services,
+		)[0]?.[1];
+
+		if (service != null) {
+			stateToSend.status = service.status;
+			if (stateToSend.status === 'Running') {
+				stateToSend.status = 'Idle';
+			}
+			stateToSend.download_progress = service.download_progress;
+		}
+		res.json(stateToSend);
+	} catch (e) {
+		res.status(500).json({
+			Data: '',
+			Error:
+				(e != null ? (e as Error).message : undefined) || e || 'Unknown error',
+		});
+	}
+});
+
+router.use(applicationManager.router);
